vsp-lib: Handle logging internally
Remove the need to pipe individual commands to the logger from the test
scripts by moving all logging code internally to vsp-lib and vsp-runner.

As an interesting side effect the removal of shell pipes from the test
scripts avoids running functions in subshells. This makes it possible
for the pipeline configuration functions to set variables that can be
used later.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>