vsp-lib: Add support for pausing and resuming yavta
When the --pause argument to vsp_runner() is specified, the
corresponding yavta instance will pause after capturing or outputting
the given number of frames and wait for a signal before resuming. The
new vsp_runner_wait() and vsp_runner_resume() functions can be used to
respectively wait for yavta to pause and instruct it to resume.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>