gen-image: Add -C argument to disable chroma averaging on output
gen-image averages consecutive chroma values when subsampling chroma
horizontally. This matches the hardware behaviour, but isn't desired to
produce test input frames in subsampled formats as otherwise the
hardware will average the values a second time in the RPF to interpolate
missing chroma components.

Add a -C argument to disable chroma averaging on output, to be used to
generate test input frames.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>