gen-image: Subsample YUV chroma on input
When the input format is YUV, subsample the chroma during RGB to YUV
conversion based on the YUV format parameters. This matches the RGB to
YUV conversion performed to generate the test input frame fed to the
RPF.

A better fix would be to add support for unpacking all RGB and YUV
formats instead of converting the .pnm reference frames.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>